[refactor] Move partial NS interfaces module into a new ns subfolder
