+++ src/ns/interfaces.ts
